+++ packages/api/lambdas/publish-granules.js
@@ -14,8 +14,27 @@ const handler = async (event) => {
 
   const promisedPublishEvents = event.Records.map(
     (record) => {
-      const granule = attr.unwrap(record.dynamodb.NewImage);
-      return publishSnsMessage(topicArn, granule);
+      const message = {};
+      const newImage = attr.unwrap(record.dynamodb.NewImage);
+      const oldImage = attr.unwrap(record.dynamodb.OldImage);
+      switch (record.eventName) {
+      case 'INSERT':
+        message.event = 'Create';
+        message.record = newImage;
+        break;
+      case 'MODIFY':
+        message.event = 'Update';
+        message.record = newImage;
+        break;
+      case 'REMOVE':
+        message.event = 'Delete';
+        message.record = oldImage;
+        message.record.deletedAt = Date.now();
+        break;
+      default:
+        break;
+      }
+      return publishSnsMessage(topicArn, message);
     }
   );